Added team id logic and refactored directory structure again. Added celebration logic to the player script
